LibAudio: Refactor out Audio::Frame to its own file
